+++ packages/ledger-icrc/src/index.canister.ts
@@ -1,16 +1,17 @@
-import { Canister, createServices } from "@dfinity/utils";
+import { createServices } from "@dfinity/utils";
 import type {
   GetTransactions,
   _SERVICE as IcrcIndexService,
 } from "../candid/icrc_index";
 import { idlFactory as certifiedIdlFactory } from "../candid/icrc_index.certified.idl";
 import { idlFactory } from "../candid/icrc_index.idl";
+import { IcrcCanister } from "./canister";
 import { toGetTransactionsArgs } from "./converters/index.converters";
 import { IndexError } from "./errors/index.errors";
 import type { IcrcLedgerCanisterOptions } from "./types/canister.options";
 import type { GetAccountTransactionsParams } from "./types/index.params";
 
-export class IcrcIndexCanister extends Canister<IcrcIndexService> {
+export class IcrcIndexCanister extends IcrcCanister<IcrcIndexService> {
   static create(options: IcrcLedgerCanisterOptions<IcrcIndexService>) {
     const { service, certifiedService, canisterId } =
       createServices<IcrcIndexService>({
@@ -30,12 +31,19 @@ export class IcrcIndexCanister extends Canister<IcrcIndexService> {
    * because the index canisters makes a call to the ledger canister to get the transaction data.
    * Index Canister only holds the transactions ids in state, not the whole transaction data.
    */
-  getTransactions = async (
-    params: GetAccountTransactionsParams,
-  ): Promise<GetTransactions> => {
+  /**
+   * Get the transactions of an account.
+   *
+   * @param {GetAccountTransactionsParams} params The parameters to get the transactions of an account.
+   * @returns {Promise<GetTransactions>} The list of transactions and further related information of the given account.
+   */
+  getTransactions = async ({
+    certified,
+    ...rest
+  }: GetAccountTransactionsParams): Promise<GetTransactions> => {
     const response = await this.caller({
-      certified: true,
-    }).get_account_transactions(toGetTransactionsArgs(params));
+      certified,
+    }).get_account_transactions(toGetTransactionsArgs(rest));
 
     if ("Err" in response) {
       throw new IndexError(response.Err.message);