+++ packages/ledger-icrc/src/ledger.canister.ts
@@ -1,5 +1,5 @@
 import type { QueryParams } from "@dfinity/utils";
-import { Canister, createServices, toNullable } from "@dfinity/utils";
+import { createServices } from "@dfinity/utils";
 import type {
   Allowance,
   BlockIndex,
@@ -8,6 +8,7 @@ import type {
 } from "../candid/icrc_ledger";
 import { idlFactory as certifiedIdlFactory } from "../candid/icrc_ledger.certified.idl";
 import { idlFactory } from "../candid/icrc_ledger.idl";
+import { IcrcCanister } from "./canister";
 import {
   toApproveArgs,
   toTransferArg,
@@ -18,13 +19,12 @@ import type { IcrcLedgerCanisterOptions } from "./types/canister.options";
 import type {
   AllowanceParams,
   ApproveParams,
-  BalanceParams,
   TransferFromParams,
   TransferParams,
 } from "./types/ledger.params";
 import type { IcrcTokenMetadataResponse } from "./types/ledger.responses";
 
-export class IcrcLedgerCanister extends Canister<IcrcLedgerService> {
+export class IcrcLedgerCanister extends IcrcCanister<IcrcLedgerService> {
   static create(options: IcrcLedgerCanisterOptions<IcrcLedgerService>) {
     const { service, certifiedService, canisterId } =
       createServices<IcrcLedgerService>({
@@ -49,18 +49,6 @@ export class IcrcLedgerCanister extends Canister<IcrcLedgerService> {
    */
   transactionFee = (params: QueryParams): Promise<Tokens> =>
     this.caller(params).icrc1_fee();
-
-  /**
-   * Returns the balance for a given account provided as owner and with optional subaccount.
-   *
-   * @param {BalanceParams} params The parameters to get the balance of an account.
-   * @returns {Promise<Tokens>} The balance of the given account.
-   */
-  balance = (params: BalanceParams): Promise<Tokens> =>
-    this.caller({ certified: params.certified }).icrc1_balance_of({
-      owner: params.owner,
-      subaccount: toNullable(params.subaccount),
-    });
 
   /**
    * Transfers tokens from the sender to the given account.