+++ packages/sns/src/sns.wrapper.spec.ts
@@ -54,6 +54,7 @@ describe("SnsWrapper", () => {
   mockIndexCanister.getTransactions.mockResolvedValue({
     transactions: [],
     oldest_tx_id: [BigInt(2)],
+    balance: 123n,
   });
 
   const snsWrapper: SnsWrapper = new SnsWrapper({